// web/src/form/questions/response/__generated__/ResponseModeParagraphQuestionResponseQuery.graphql.ts
export type ResponseModeParagraphQuestionResponseQuery$variables = {
  formInstanceID: string;
  questionID: string;
};
export type ResponseModeParagraphQuestionResponseQuery$data = {
  readonly questionResponses: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly value: string;
      } | null | undefined;
    } | null | undefined> | null | undefined;
  };
};
export type ResponseModeParagraphQuestionResponseQuery = {
  response: ResponseModeParagraphQuestionResponseQuery$data;
  variables: ResponseModeParagraphQuestionResponseQuery$variables;
};